+++ packages/blade/src/components/Button/IconButton/IconButton.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import StyledIconButton from './StyledIconButton';
 import type { IconComponent } from '~components/Icons';
 import type { BladeElementRef } from '~src/hooks/useBladeInnerRef';
-import { useBladeInnerRef } from '~src/hooks/useBladeInnerRef';
+import type { BladeCommonEvents } from '~components/types';
 
 type IconButtonProps = {
   /**
@@ -31,25 +31,46 @@ type IconButtonProps = {
    * Sets aria-label to help users know what the action does, eg 'Dismiss alert'
    */
   accessibilityLabel: string;
-};
+} & BladeCommonEvents;
 
 /**
  * Component for making clickable icons with transparent background.
  * For other cases please use `Button` component with `icon` prop.
  */
 const _IconButton: React.ForwardRefRenderFunction<BladeElementRef, IconButtonProps> = (
-  { icon, onClick, size = 'medium', contrast = 'low', accessibilityLabel },
+  {
+    icon,
+    onClick,
+    size = 'medium',
+    contrast = 'low',
+    accessibilityLabel,
+    onBlur,
+    onFocus,
+    onMouseLeave,
+    onMouseMove,
+    onPointerDown,
+    onPointerEnter,
+    onTouchEnd,
+    onTouchStart,
+  },
   ref,
 ) => {
-  const innerRef = useBladeInnerRef(ref);
   return (
     <StyledIconButton
-      ref={innerRef as any}
+      ref={ref as any}
       onClick={onClick}
       contrast={contrast}
       size={size}
       icon={icon}
       accessibilityLabel={accessibilityLabel}
+      onBlur={onBlur}
+      onFocus={onFocus}
+      onMouseLeave={onMouseLeave}
+      onMouseMove={onMouseMove}
+      onPointerDown={onPointerDown}
+      onPointerEnter={onPointerEnter}
+      onTouchEnd={onTouchEnd}
+      onTouchStart={onTouchStart}
     />
   );
 };