+++ packages/blade/src/hooks/types.ts
@@ -3,7 +3,9 @@ import type { View } from 'react-native';
 import type { Platform } from '~utils';
 
 type BladeElementRef = Platform.Select<{
-  web: Pick<HTMLElement, 'focus' | 'scrollIntoView' | 'clientHeight'> | Pick<View, 'focus'>;
+  web:
+    | Pick<HTMLElement, 'focus' | 'scrollIntoView' | 'getBoundingClientRect' | 'clientHeight'>
+    | Pick<View, 'focus'>;
   native: React.MutableRefObject<any>;
 }>;