+++ packages/blade/src/hooks/useBladeInnerRef.web.ts
@@ -25,6 +25,7 @@ const useBladeInnerRef = (
       return {
         focus: (opts) => (handlers?.onFocus ? handlers.onFocus(opts) : element?.focus?.(opts)),
         scrollIntoView: (opts) => element?.scrollIntoView?.(opts),
+        getBoundingClientRect: () => element?.getBoundingClientRect(),
         clientHeight: element.clientHeight,
       };
     },