+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @razorpay/blade
 
+## 8.9.0
+
+### Minor Changes
+
+- d62a86d4: feat(blade): added tooltip component
+
+  Blade now has a peer dependency on [FloatingUI](https://floating-ui.com/), make sure to install it.
+
+  For web:
+
+  ```
+  yarn add @floating-ui/react
+  ```
+
+  For native:
+
+  ```
+  yarn add @floating-ui/react-native
+  ```
+
+## 8.8.4
+
+### Patch Changes
+
+- 333d0e7e: fix(DropdownOverlay): fix incorrect minWidth condition
+
 ## 8.8.3
 
 ### Patch Changes